Fix broken plugin

+++ backend/src/main/java/com/example/spring_project/controller/InfoController.java
@@ -1,4 +1,5 @@
 package com.example.spring_project.controller;
+
 import com.example.spring_project.model.Info;
 import com.example.spring_project.service.InfoService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -6,9 +7,6 @@ import org.springframework.web.bind.annotation.*;
 import com.example.spring_project.model.Info_DTO;
 
 import java.util.List;
-
-
-
 
 @RestController
 @RequestMapping("/api/info")
@@ -21,12 +19,14 @@ public class InfoController {
     public List<Info> getAllInfo() {
         return infoService.getAllInfo();
     }
+
     @GetMapping("/{id}")
-    public List<Info> getInfoByLessionID(@PathVariable int id) { return infoService.getInfoByLessionId(id);}
+    public List<Info> getInfoByLessionID(@PathVariable Long id) {
+        return infoService.getInfoByLessionId(id);
+    }
 
     @PostMapping
     public Info addInfo(@RequestBody Info_DTO info) {
         return infoService.addInfo(info);
     }
 }
-
